requirements.txt
Django==3.1.7
djangorestframework==3.12.2
gunicorn==20.0.4
whitenoise==5.2.0
dj-static==0.0.6
decouple==0.0.7
Pillow==8.1.2
psycopg2==2.8.6
